move typings to types folder & nodenext compatible
